Moved CalculateLevels and HabitList into services file
